Add the `firstWillUpdate` lifecycle method and improve initial render perf by reducing the Total Blocking Time (#4)
* Remove the defer* properties of the component decorator in favor of a built-in scheduler

The deferInitialRender and deferUpdate properties of the Component decorator has been removed, since we are now implementing a scheduler in the Component decorator to provide a more automatic way for deferring updates to improve the render performance.

* Partition in more batches the update life cycle

* Add weights for the type of update

* Add `beforeFirstUpdate` lifecycle method

Invoked before the first `willUpdate` of the component. Subsequent renders will not call this method, as it under the hood is implemented by monkey-patching the `willUpdate` method and checking for the `this.hasUpdated` property.

* Rename "beforeFirstUpdate" to "firstWillUpdate"

* Improve docs

* Move globalStyleSheet implementation to the global-sttylesheets.ts file

* Add e2e test for the scheduler

* Fix scheduler issues

* Add more e2e tests

* Improve e2e tests

* Minor tweak of the weights for the scheduler

* Delete old properties

* Fix tests

---------

Co-authored-by: ncamera <nicolas.camera+globant@globant.com>